moved each component to its directory, added __test__ directory inside and created test file for each one
