--- src/components/PayPalButton.vue
@@ -1,16 +1,0 @@
-<template>
-  <div>
-    <button class="w-72  h-14 border-2 bg-black text-white"> Pagar con Paypal</button>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "PayPalButton",
-
-};
-</script>
-
-<style>
-
-</style>
