+++ src/components/StoreBadge/StoreBadge.vue
