+++ src/components/StoreFeaturedItem/StoreFeaturedItem.vue
